take out circular queue changes.
