+++ graphbolt/src/cache_policy.h
@@ -24,10 +24,65 @@
 #include <torch/custom_class.h>
 #include <torch/torch.h>
 
-#include "./circular_queue.h"
-
 namespace graphbolt {
 namespace storage {
+
+template <typename T>
+struct CircularQueue {
+  CircularQueue(const int64_t capacity)
+      : tail_(0),
+        head_(0),
+        // + 1 is needed to be able to differentiate empty and full states.
+        capacity_(capacity + 1),
+        data_{new T[capacity + 1]} {}
+
+  CircularQueue() = default;
+
+  T* Push(const T& x) {
+    auto insert_ptr = &data_[PostIncrement(tail_)];
+    *insert_ptr = x;
+    return insert_ptr;
+  }
+
+  T Pop() { return data_[PostIncrement(head_)]; }
+
+  T& Front() const { return data_[head_]; }
+
+  bool IsFull() const {
+    const auto diff = tail_ + 1 - head_;
+    return diff == 0 || diff == capacity_;
+  }
+
+  auto Size() const {
+    auto diff = tail_ - head_;
+    if (diff < 0) diff += capacity_;
+    return diff;
+  }
+
+  friend std::ostream& operator<<(
+      std::ostream& os, const CircularQueue& queue) {
+    for (auto i = queue.head_; i != queue.tail_; queue.PostIncrement(i)) {
+      os << queue.data_[i] << ", ";
+    }
+    return os << "\n";
+  }
+
+  bool IsEmpty() const { return tail_ == head_; }
+
+  auto Capacity() const { return capacity_ - 1; }
+
+ private:
+  int64_t PostIncrement(int64_t& i) const {
+    const auto ret = i++;
+    if (i >= capacity_) i -= capacity_;
+    return ret;
+  }
+
+  int64_t tail_;
+  int64_t head_;
+  int64_t capacity_;
+  std::unique_ptr<T[]> data_;
+};
 
 struct CacheKey {
   CacheKey(int64_t key, int64_t position)
